requirements.txt
APScheduler==3.6.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
google-auth==1.16.1
idna==2.9
kubernetes==24.2.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.9
websocket-client==0.57.0